fix: adapt new main test to FlowResult API; rename dadb logcat helpers
- IntegrationTest: main's new 'optional launchApp ... warned, not failed' test
  used runFlow's old Boolean return; this branch returns FlowResult, so assert
  result.success (surfaced only at test-compile after the main merge).
- Rename DadbExtensions.kt -> LogcatExtensions.kt and move from the
  maestro.android.dadb package into maestro.android: the helpers now extend
  AndroidDeviceConnection (not raw Dadb), so the 'dadb' name was stale.
