+++ maestro-test/src/test/kotlin/maestro/test/IntegrationTest.kt
@@ -4924,7 +4924,7 @@ class IntegrationTest {
         var onCommandFailedCalled = false
 
         Maestro(driver).use { maestro ->
-            val success = runBlocking {
+            val result = runBlocking {
                 Orchestra(
                     maestro,
                     lookupTimeoutMs = 0L,
@@ -4936,7 +4936,7 @@ class IntegrationTest {
                     },
                 ).runFlow(commands)
             }
-            assertThat(success).isTrue()
+            assertThat(result.success).isTrue()
         }
         assertThat(onCommandWarnedCalled).isTrue()
         assertThat(onCommandFailedCalled).isFalse()
